requirements.txt
opendartreader>=0.2.0
finance-datareader>=0.9.0
pandas>=2.0.0
streamlit>=1.30.0
plotly>=5.18.0
python-dotenv>=1.0.0
openpyxl>=3.1.0
